# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ICTU AI Judge API
  description: RESTful API spec for ICTU AI Judge Platform (FastAPI Backend)
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development Server

# Bảo mật: Yêu cầu JWT Token truyền ngầm qua HttpOnly Cookie thay vì Header Bearer.
security:
  - cookieAuth: []
paths:
...
  /challenges:
    get:
      summary: Lấy danh sách Bài thi
      tags:
        - Challenges
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, ARCHIVED]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Danh sách bài thi (Phân trang).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Challenge'
                  
    post:
      summary: Tạo Bài thi mới (Admin)
      tags:
        - Challenges
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeCreate'
      responses:
        '201':
          description: Tạo cấu trúc thông tin JSON thành công.
  /challenges/{id}/upload-secrets:
    post:
      summary: Upload File Đáp án và Script chấm điểm (Dành cho Admin)
      tags:
        - Challenges
      description: Upload File Ground Truth và Custom Metric để lưu an toàn trên Storage ngầm, thí sinh không thể truy cập.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                ground_truth_csv:
                  type: string
                  format: binary
                  description: File đáp án bắt buộc chứa cột 'Usage'
                metric_script_py:
                  type: string
                  format: binary
                  description: (Tùy chọn) Script python chấm điểm thủ công
                public_test_split_ratio:
                  type: integer
                  description: Tỉ lệ % tập Public (0-100)
                  default: 30
      responses:
        '200':
          description: Upload thành công.
  /challenges/{id}:
    get:
      summary: Chi tiết Bài thi
      tags:
        - Challenges
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Thông tin chi tiết.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Challenge'
    patch:
      summary: Cập nhật Bài thi (Admin)
      tags:
        - Challenges
      description: Cập nhật thông tin bài thi (mô tả, thời gian, luật thi). Bị khóa nếu bài đã có Submission.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeUpdate'
      responses:
        '200':
          description: Cập nhật thành công.
        '403':
          description: Không có quyền (không phải Admin hoặc bài thi đã lock).
    delete:
      summary: Xóa mềm Bài thi (Admin - Soft Delete)
      tags:
        - Challenges
      description: Đặt `deleted_at` = now(), không xóa cứng khỏi Database.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Xóa mềm thành công.
        '403':
          description: Không có quyền hoặc bài thi đang diễn ra.
  /challenges/{id}/enroll:
    post:
      summary: Ghi danh (Tự động tạo Team of 1)
      tags:
        - Challenges
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Ghi danh thành công, trả về Team ID.
  /challenges/{id}/participants:
    get:
      summary: Xem danh sách Whitelist (Admin)
      tags:
        - Challenges
        - Admin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Danh sách sinh viên được cấp quyền tham gia Competition.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Participant'
    post:
      summary: Thêm Whitelist Participant (Admin)
      tags:
        - Challenges
        - Admin
      description: Thêm một hoặc nhiều user vào danh sách được phép tham gia Competition.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  description: Danh sách User ID cần thêm vào whitelist
      responses:
        '201':
          description: Thêm whitelist thành công.
        '403':
          description: Không có quyền Admin.
  /challenges/{id}/participants/by-identifiers:
    post:
      summary: Thêm Whitelist Participant bằng định danh (Admin)
      tags:
        - Challenges
        - Admin
      description: Thêm một hoặc nhiều user vào danh sách được phép tham gia Competition bằng Email, MSSV, hoặc UUID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifiers:
                  type: array
                  items:
                    type: string
                  description: Danh sách định danh (Email, MSSV, UUID)
      responses:
        '200':
          description: Thêm whitelist thành công, trả về số lượng thành công và các lỗi.
        '403':
          description: Không có quyền Admin.
        '422':
          description: Validation Error (Ví dụ trường identifiers trống).

  # ==========================================
  # TEAM ENDPOINTS
  # ==========================================
...
  /challenges/{id}/submissions:
    get:
      summary: Lịch sử nộp bài của Đội
      tags:
        - Submissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Danh sách lượt nộp (Phân trang).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationMeta'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Submission'

    post:
      summary: Upload file Nộp bài
      tags:
        - Submissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File CSV dự đoán
      responses:
        '201':
          description: File hợp lệ, bắt đầu chấm điểm ngầm (Trạng thái PENDING).
        '429':
          description: Vi phạm Rate Limit (Xin chờ X phút).
        '409':
          description: Trùng Hash MD5 (Spam, file này vừa nộp rồi).
...
  /admin/challenges/{challenge_id}/submissions:
    get:
      summary: Lấy danh sách tất cả bài nộp của cuộc thi (Admin)
      tags:
        - Admin
      parameters:
        - name: challenge_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Danh sách bài nộp thành công.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionList'
        '403':
          description: Không có quyền Admin.
...
  /challenges/{challenge_id}/solutions:
    get:
      tags: [Solutions]
      summary: L?y danh s�ch Solutions c?a m?t b�i thi
      description: Tr? v? danh s�ch gi?i ph�p/notebook d� du?c chia s?, k�m presigned URL d? download v� author_name.
      security: []
      parameters:
        - name: challenge_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Danh s�ch Solutions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolutionList"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags: [Solutions]
      summary: Chia s? gi?i ph�p (Upload notebook)
      description: |
        Upload file Jupyter Notebook (.ipynb) l�n MinIO v� t?o b?n ghi Solution trong DB.
        Ch? kh? d?ng khi Challenge dang PUBLISHED.
        Y�u c?u dang nh?p.
      security:
        - cookieAuth: []
      parameters:
        - name: challenge_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, content, file]
              properties:
                title:
                  type: string
                  description: Ti�u d? solution
                  example: "Gi?i ph�p LGBM d?t 0.95 Accuracy"
                content:
                  type: string
                  description: M� t?/gi?i th�ch gi?i ph�p (Markdown)
                  example: "## Approach\n\nD�ng LightGBM v?i feature engineering..."
                file:
                  type: string
                  format: binary
                  description: File Jupyter Notebook (.ipynb)
      responses:
        "201":
          description: Solution d� du?c t?o th�nh c�ng
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: File kh�ng ph?i .ipynb ho?c challenge kh�ng PUBLISHED
  /challenges/{challenge_id}/solutions/{solution_id}/upvote:
    post:
      tags: [Solutions]
      summary: Upvote m?t Solution
      description: |
        Tang upvote count l�n +1. M?i user ch? du?c upvote m?i solution 1 l?n.
        Double-vote du?c ngan b?ng UNIQUE constraint tr�n DB (atomic, race-condition-safe).
        Y�u c?u dang nh?p.
      security:
        - cookieAuth: []
      parameters:
        - name: challenge_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: solution_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Solution sau khi upvote (upvotes + 1)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: User d� upvote solution n�y r?i (double-vote)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

          description: T�n hi?n th? c?a t�c gi?
          example: "Nguy?n Van A"
        title:
          type: string
          example: "Gi?i ph�p LGBM d?t 0.95 Accuracy"
        content:
          type: string
          description: M� t? gi?i ph�p (Markdown)
        notebook_url:
          type: string
          description: Presigned URL d? download Jupyter Notebook (TTL 1 gi?)
          example: "http://localhost:9000/ictu-ai-judge-bucket/solutions/..."
        upvotes:
          type: integer
          example: 42
        created_at:
          type: string
          format: date-time

    SolutionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Solution"
        total:
          type: integer
        page:
          type: integer
          default: 1
        size:
          type: integer
          default: 100

  # ==========================================
  # Feature: User Profile (Issue #30)
  # ==========================================
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT Token được lưu an toàn trong HttpOnly Cookie.
  parameters:
    PageParam:
      name: page
      in: query
      description: Vị trí trang hiện tại
      schema:
        type: integer
        default: 1
    SizeParam:
      name: size
      in: query
      description: Số lượng item trên một trang
      schema:
        type: integer
        default: 20
  schemas:
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 1500
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 75
...
    Challenge:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        is_public:
          type: boolean
          description: true nếu là PUBLIC, false nếu là COMPETITION
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        max_team_size:
          type: integer
          default: 1
        metric_direction:
          type: string
        dataset_url:
          type: string
          format: uri
    ChallengeCreate:
      type: object
      required:
        - title
        - type
        - start_time
        - metric_name
        - metric_direction
      properties:
        title:
          type: string
          example: "Phân loại ảnh CIFAR-10"
        description:
          type: string
          description: Mô tả bài toán (hỗ trợ Markdown)
        type:
          type: string
          enum: [PUBLIC, COMPETITION]
          description: Loại bài thi
        status:
          type: string
          enum: [DRAFT, PUBLISHED]
          default: DRAFT
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        team_lock_deadline:
          type: string
          format: date-time
          description: Hạn chật khóa đội (trước end_time)
        rate_limit_minutes:
          type: integer
          default: 10
          description: Khoảng cách tối thiểu giữa 2 lần nộp (phút)
        max_file_size_mb:
          type: integer
          default: 50
          description: Giới hạn dung lượng file upload (MB)
        max_team_size:
          type: integer
          default: 1
          description: Kích thước tối đa của một đội thi
        metric_name:
          type: string
          example: "Accuracy"
        metric_direction:
          type: string
          enum: [HIGHER_IS_BETTER, LOWER_IS_BETTER]
        dataset_url:
          type: string
          format: uri
          description: Link Google Drive chứa dataset cho sinh viên tải về
    ChallengeUpdate:
      type: object
      description: Tất cả trường đều là tùy chọn (Partial Update). Bị khóa nếu có Submission.
      properties:
        title:
          type: string
        description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
          nullable: true
        team_lock_deadline:
          type: string
          format: date-time
        rate_limit_minutes:
          type: integer
        max_file_size_mb:
          type: integer
        dataset_url:
          type: string
          format: uri
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
    Submission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        public_score:
          type: number
          format: double
        status:
          type: string
          enum: [PENDING, PROCESSING, SUCCESS, FAILED]
        is_selected_for_private:
          type: boolean
        file_size_bytes:
          type: integer
          description: Kích thước file thực tế
        execution_time_ms:
          type: integer
          description: Thời gian Worker chấm bài (ms)
        error_message:
          type: string
          nullable: true
        submitted_at:
          type: string
          format: date-time
...
    Participant:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
        full_name:
          type: string
        is_approved:
          type: boolean
        joined_at:
          type: string
          format: date-time
    Solution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        challenge_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        author_name:
          type: string
...